[finance_report_vision] Split market_data.py (1621 lines) into a package
Behaviour-preserving split of the market-data sync service. The public surface
(sync_fx_rates / sync_stock_prices / ensure_market_data_fresh /
get_market_data_status / resolve_missing_fx_rate + value objects) is preserved
via __init__ re-exports.

Package:
- _base.py: constants, quanta, provider URLs, logger.
- _types.py: observations, validated series, sync result, scope/freshness, stored rows.
- _util.py: normalization, quantization, scope keys, date windows, selection.
- _store.py: DB load/persist of FX rates, stock prices, sync state.
- _providers.py: Yahoo/Stooq HTTP fetch + parsing + validated fetch + sync specs.
- service.py: public sync orchestration.

Test decoupling (the reason this needed care): the tests reach into ~50 internal
symbols via `from src.services import market_data` + `market_data.X`, and patch
~20 provider/store helpers with `setattr(market_data, "X", fake)`. After the
split each helper lives in a submodule and its callers bind it there, so a
package-level patch no longer reaches them. Two mechanical, principled fixes:
1. __init__ re-exports the full submodule surface (fixes every direct call/
   construct access — `market_data.X` resolves again).
2. Each `setattr(market_data, "X")` is repointed to the submodule that DEFINES
   X (`market_data._providers.X` / `market_data._store.X`). This works because
   every patched fetch helper is called within _providers (the validated→derived
   →yahoo/stooq chain, incl. the _MarketSyncSpec `fetch_series` lambdas that look
   up by name at call time) and every patched store helper within _store.
   The one cross-module caller — service.resolve_missing_fx_rate → provider
   fetch — now calls it module-qualified (`_providers._fetch_*`) so the single
   _providers patch target stays consistent.

Guards: the hard-coded `market_data.py` paths in test_base_package_migration_
cleanup / test_decimal_boundary_policy / test_unit_price_adoption are repointed
to the package dir (dir-aware `_read`). Doc prose refs updated to the package.

Verification: 100 market_data + transaction_boundaries tests pass; the 3 path
guards pass; full collection clean (2673); ruff check/format clean under both
the apps/backend and repo-root configs.
