+++ apps/backend/src/money/rounding.py
@@ -11,7 +11,7 @@ cross-language standard, #1171); ``src.utils`` re-exports ``to_money``/
 ``common/`` is not shipped into the backend image.
 
 Out of scope (deliberately use their own quantization):
-- FX rates and security prices: 6 dp (``services/market_data.py``).
+- FX rates and security prices: 6 dp (``services/market_data/``).
 - Share quantities: 6 dp (``services/investment_accounting.py``).
 - Percentages / performance ratios (XIRR, TWR, MWR, allocation %): not currency.
 """
