+++ apps/backend/src/services/market_data/__init__.py
@@ -1,0 +1,193 @@
+"""Market data sync service (package)."""
+
+import httpx  # noqa: F401
+
+from src.services.market_data._providers import (
+    _FX_SYNC_SPEC,
+    _STOCK_SYNC_SPEC,
+    _fetch_provider_response,
+    _fetch_stooq_fx_rate,
+    _fetch_stooq_fx_rate_series,
+    _fetch_stooq_stock_price,
+    _fetch_stooq_stock_price_series,
+    _fetch_validated_fx_rate,
+    _fetch_validated_fx_rate_series,
+    _fetch_validated_fx_rate_series_for_scope,
+    _fetch_validated_stock_price,
+    _fetch_validated_stock_price_series,
+    _fetch_yahoo_fx_rate,
+    _fetch_yahoo_fx_rate_series,
+    _fetch_yahoo_or_derived_fx_rate,
+    _fetch_yahoo_or_derived_fx_rate_series,
+    _fetch_yahoo_stock_price,
+    _fetch_yahoo_stock_price_series,
+    _parse_stooq_close_rows,
+    _parse_stooq_fx_csv_series,
+    _parse_stooq_stock_csv_series,
+    _parse_yahoo_close_rows,
+    _parse_yahoo_fx_response_series,
+    _parse_yahoo_stock_response_series,
+    _stooq_daily_params,
+    _yahoo_chart_params,
+)
+from src.services.market_data._store import (
+    _active_stock_symbols,
+    _derive_from_bridge_rates,
+    _fallback_last_success_at,
+    _is_sync_scope_fresh,
+    _latest_fx_rate_date,
+    _latest_fx_rate_date_for_scope,
+    _latest_observation_date,
+    _latest_observation_date_on_or_before,
+    _latest_stock_price_date,
+    _load_stored_direct_or_inverse,
+    _load_stored_rate,
+    _load_stored_rate_on_date,
+    _load_stored_stock_price,
+    _load_stored_stock_price_on_date,
+    _load_sync_state,
+    _observed_fx_pairs,
+    _persist_fx_rate,
+    _persist_stock_price,
+    _stored_fx_rate_dates,
+    _stored_fx_rate_dates_for_scope,
+    _stored_stock_price_dates,
+    _sync_scope_status,
+    _upsert_sync_state,
+)
+from src.services.market_data._types import (
+    FxRateObservation,
+    MarketDataFreshnessResult,
+    MarketDataScopeStatus,
+    MarketDataSyncResult,
+    MarketObservation,
+    ProviderDisagreement,
+    StockPriceObservation,
+    ValidatedMarketObservation,
+    ValidatedMarketObservationSeries,
+    _MarketSyncSpec,
+    _StoredFxRate,
+    _StoredStockPrice,
+)
+from src.services.market_data._util import (
+    _date_to_epoch,
+    _default_start_date,
+    _fx_scope,
+    _incremental_start,
+    _iter_dates,
+    _looks_like_ticker,
+    _normalize_currency,
+    _normalize_symbol,
+    _normalize_utc,
+    _observation_date,
+    _observations_by_date,
+    _parse_fx_pair,
+    _parse_sync_fx_scope,
+    _quantize_price,
+    _quantize_rate,
+    _relative_difference,
+    _select_validated_observation,
+    _select_validated_observation_series,
+    _stock_scope,
+    _stooq_fx_symbol,
+    _stooq_stock_symbol,
+)
+from src.services.market_data.service import (
+    _sync_market_observation_series,
+    ensure_market_data_fresh,
+    get_market_data_status,
+    resolve_missing_fx_rate,
+    sync_fx_rates,
+    sync_stock_prices,
+)
+
+__all__ = [
+    "FxRateObservation",
+    "MarketDataFreshnessResult",
+    "MarketDataScopeStatus",
+    "MarketDataSyncResult",
+    "MarketObservation",
+    "ProviderDisagreement",
+    "StockPriceObservation",
+    "ValidatedMarketObservation",
+    "ValidatedMarketObservationSeries",
+    "_FX_SYNC_SPEC",
+    "_MarketSyncSpec",
+    "_STOCK_SYNC_SPEC",
+    "_StoredFxRate",
+    "_StoredStockPrice",
+    "_active_stock_symbols",
+    "_date_to_epoch",
+    "_default_start_date",
+    "_derive_from_bridge_rates",
+    "_fallback_last_success_at",
+    "_fetch_provider_response",
+    "_fetch_stooq_fx_rate",
+    "_fetch_stooq_fx_rate_series",
+    "_fetch_stooq_stock_price",
+    "_fetch_stooq_stock_price_series",
+    "_fetch_validated_fx_rate",
+    "_fetch_validated_fx_rate_series",
+    "_fetch_validated_fx_rate_series_for_scope",
+    "_fetch_validated_stock_price",
+    "_fetch_validated_stock_price_series",
+    "_fetch_yahoo_fx_rate",
+    "_fetch_yahoo_fx_rate_series",
+    "_fetch_yahoo_or_derived_fx_rate",
+    "_fetch_yahoo_or_derived_fx_rate_series",
+    "_fetch_yahoo_stock_price",
+    "_fetch_yahoo_stock_price_series",
+    "_fx_scope",
+    "_incremental_start",
+    "_is_sync_scope_fresh",
+    "_iter_dates",
+    "_latest_fx_rate_date",
+    "_latest_fx_rate_date_for_scope",
+    "_latest_observation_date",
+    "_latest_observation_date_on_or_before",
+    "_latest_stock_price_date",
+    "_load_stored_direct_or_inverse",
+    "_load_stored_rate",
+    "_load_stored_rate_on_date",
+    "_load_stored_stock_price",
+    "_load_stored_stock_price_on_date",
+    "_load_sync_state",
+    "_looks_like_ticker",
+    "_normalize_currency",
+    "_normalize_symbol",
+    "_normalize_utc",
+    "_observation_date",
+    "_observations_by_date",
+    "_observed_fx_pairs",
+    "_parse_fx_pair",
+    "_parse_stooq_close_rows",
+    "_parse_stooq_fx_csv_series",
+    "_parse_stooq_stock_csv_series",
+    "_parse_sync_fx_scope",
+    "_parse_yahoo_close_rows",
+    "_parse_yahoo_fx_response_series",
+    "_parse_yahoo_stock_response_series",
+    "_persist_fx_rate",
+    "_persist_stock_price",
+    "_quantize_price",
+    "_quantize_rate",
+    "_relative_difference",
+    "_select_validated_observation",
+    "_select_validated_observation_series",
+    "_stock_scope",
+    "_stooq_daily_params",
+    "_stooq_fx_symbol",
+    "_stooq_stock_symbol",
+    "_stored_fx_rate_dates",
+    "_stored_fx_rate_dates_for_scope",
+    "_stored_stock_price_dates",
+    "_sync_market_observation_series",
+    "_sync_scope_status",
+    "_upsert_sync_state",
+    "_yahoo_chart_params",
+    "ensure_market_data_fresh",
+    "get_market_data_status",
+    "resolve_missing_fx_rate",
+    "sync_fx_rates",
+    "sync_stock_prices",
+]
